Types of EDP functions fixed

+++ Include/EDP/LocalHamiltonian.hpp
@@ -8,12 +8,12 @@ namespace internal
 {
 	inline uint32_t ipow(uint32_t base, uint32_t exp)
 	{
-		int result = 1u;
+		uint32_t result = 1U;
 		while (exp)
 		{
-			if (exp & 1)
+			if (exp & 1U)
 				result *= base;
-			exp >>= 1;
+			exp >>= 1U;
 			base *= base;
 		}
 
@@ -24,14 +24,14 @@ namespace internal
 template<typename T>
 struct TwoSiteTerm
 {
-	std::pair<int, int> sites;
+	std::pair<uint32_t, uint32_t> sites;
 	Eigen::SparseMatrix<T> m;
 
-	TwoSiteTerm(const std::pair<int,int>& p1, const Eigen::SparseMatrix<T>& p2)
+	TwoSiteTerm(const std::pair<uint32_t, uint32_t>& p1, const Eigen::SparseMatrix<T>& p2)
 		: sites(p1), m(p2)
 	{
 	}
-	TwoSiteTerm(std::pair<int,int>&& p1, Eigen::SparseMatrix<T>&& p2)
+	TwoSiteTerm(std::pair<uint32_t, uint32_t>&& p1, Eigen::SparseMatrix<T>&& p2)
 		: sites(p1), m(p2)
 	{
 	}
@@ -40,14 +40,15 @@ struct TwoSiteTerm
 template<typename T>
 struct OneSiteTerm
 {
-	int site;
+	uint32_t site;
 	Eigen::SparseMatrix<T> m;
 
-	OneSiteTerm(int p1, const Eigen::SparseMatrix<T>& p2)
+	OneSiteTerm(uint32_t p1, const Eigen::SparseMatrix<T>& p2)
 		: site(p1), m(p2)
 	{
 	}
-	OneSiteTerm(int p1, Eigen::SparseMatrix<T>&& p2)
+
+	OneSiteTerm(uint32_t p1, Eigen::SparseMatrix<T>&& p2)
 		: site(p1), m(p2)
 	{
 	}
@@ -57,20 +58,20 @@ template<typename T>
 class LocalHamiltonian
 {
 private:
-	int numSites_;
-	int d_; //local Hibert dimension
+	uint32_t numSites_;
+	uint64_t d_; //local Hibert dimension
 
 	std::vector<TwoSiteTerm<T> > twoSiteTerms_;
 	std::vector<OneSiteTerm<T> > oneSiteTerms_;
 
-	int swapBaseD(int idx, int pos, int val) const
+	uint32_t swapBaseD(uint32_t idx, uint32_t pos, uint32_t val) const
 	{
-		int b = internal::ipow(d_, pos);
-		int upper = (idx / (b*d_))*d_ + val;
+		uint32_t b = internal::ipow(d_, pos);
+		uint32_t upper = (idx / (b*d_))*d_ + val;
 		return upper*b + (idx % b);
 	}
 public:
-	LocalHamiltonian(int numSites, int d)
+	LocalHamiltonian(uint32_t numSites, uint32_t d)
 		: numSites_{numSites}, d_{d}
 	{
 	}
@@ -85,7 +86,7 @@ public:
 		std::vector<TwoSiteTerm<T> >().swap(twoSiteTerms_);
 		std::vector<OneSiteTerm<T> >().swap(oneSiteTerms_);
 	}
-	int getNumSites() const
+	uint32_t getNumSites() const
 	{
 		return numSites_;
 	}
@@ -101,7 +102,7 @@ public:
 		m.makeCompressed();
 		twoSiteTerms_.emplace_back(site, std::move(m));
 	}
-	void addOneSiteTerm(int site, Eigen::SparseMatrix<T> m)
+	void addOneSiteTerm(uint32_t site, Eigen::SparseMatrix<T> m)
 	{
 		m.makeCompressed();
 		oneSiteTerms_.emplace_back(site, std::move(m));
@@ -126,8 +127,8 @@ std::map<uint32_t, T> edp::LocalHamiltonian<T>::getCol(uint32_t n) const
 		auto col = b*d_ + a;
 		for(typename SparseMatrix<T>::InnerIterator it(twoSiteTerm.m, col); it; ++it)
 		{
-			int r = it.row();
-			int t = n;
+			uint32_t r = it.row();
+			uint32_t t = n;
 			t = swapBaseD(t, twoSiteTerm.sites.first, r%d_);
 			t = swapBaseD(t, twoSiteTerm.sites.second, r/d_);
 			m[t] += it.value();
@@ -135,14 +136,14 @@ std::map<uint32_t, T> edp::LocalHamiltonian<T>::getCol(uint32_t n) const
 	}
 	for(auto& oneSiteTerm: oneSiteTerms_)
 	{
-		int a = (n/ipow(d_,oneSiteTerm.site))%d_;
+		uint32_t a = (n/ipow(d_,oneSiteTerm.site))%d_;
 		//auto col = oneSiteTerm.m.col(a);
 		auto col = a;
 
 		for(typename SparseMatrix<T>::InnerIterator it(oneSiteTerm.m, col); it; ++it)
 		{
-			int r = it.row();
-			int t = n;
+			uint32_t r = it.row();
+			uint32_t t = n;
 			t = swapBaseD(t, oneSiteTerm.site, r);
 			m[t] += it.value();
 		}
